split api_communication into different files
